Tema07 - Apis REST/Clase21_21-11-2025/api-libros-flaskrestful/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-restful==0.3.10
psycopg2-binary==2.9.11
python-dotenv==1.2.1